chore: add apiStrict-only test to CI MONGOSH-756 (#943)
Run the shell-api and cli-repl test suites with the `strict` and
`deprecationErrors` versioned API options, and skip tests where
this does not apply.

+++ packages/cli-repl/src/cli-repl.spec.ts
@@ -991,7 +991,10 @@ describe('CliRepl', () => {
       context('for server >= 4.1', () => {
         skipIfServerVersion(testServer, '< 4.1');
 
-        it('terminates operations on the server side', async() => {
+        it('terminates operations on the server side', async function() {
+          if (process.env.MONGOSH_TEST_FORCE_API_STRICT) {
+            return this.skip(); // $currentOp is unversioned
+          }
           input.write('db.ctrlc.find({ $where: \'while(true) { /* loop1 */ }\' })\n');
           await delay(100);
           process.kill(process.pid, 'SIGINT');
@@ -1161,7 +1164,10 @@ describe('CliRepl', () => {
         expect(output).not.to.include('listCollections requires authentication');
       });
 
-      it(`${wantWatch ? 'completes' : 'does not complete'} the watch method`, async() => {
+      it(`${wantWatch ? 'completes' : 'does not complete'} the watch method`, async function() {
+        if (process.env.MONGOSH_TEST_FORCE_API_STRICT) {
+          return this.skip();
+        }
         output = '';
         input.write('db.wat');
         await tabtab();
@@ -1173,7 +1179,10 @@ describe('CliRepl', () => {
         }
       });
 
-      it('completes the version method', async() => {
+      it('completes the version method', async function() {
+        if (process.env.MONGOSH_TEST_FORCE_API_STRICT) {
+          return this.skip();
+        }
         output = '';
         input.write('db.vers');
         await tabtab();
@@ -1181,7 +1190,10 @@ describe('CliRepl', () => {
         expect(output).to.include('db.version');
       });
 
-      it(`${wantShardDistribution ? 'completes' : 'does not complete'} the getShardDistribution method`, async() => {
+      it(`${wantShardDistribution ? 'completes' : 'does not complete'} the getShardDistribution method`, async function() {
+        if (process.env.MONGOSH_TEST_FORCE_API_STRICT) {
+          return this.skip();
+        }
         output = '';
         input.write('db.coll.getShardDis');
         await tabtab();
